// mobile/src/gen/types/JobOut.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { JobStatus } from "./JobStatus.ts";

/**
 * JobOut
 */
export type JobOut = {
  /**
   * @type integer
   */
  id: number;
  /**
   * @type integer
   */
  client_id: number;
  /**
   * @type string
   */
  client_name: string;
  series_id: number | null;
  /**
   * @type string, date
   */
  scheduled_date: string;
  scheduled_time: string | null;
  /**
   * @type integer
   */
  duration_minutes: number;
  /**
   * @type string
   */
  status: JobStatus;
  /**
   * @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
   * @type string
   */
  price: string;
  /**
   * @type boolean
   */
  paid: boolean;
  /**
   * @type string
   */
  notes: string;
  route_order: number | null;
};